Basic confirmation request on delete

+++ src/mechanics/formMechanic.ts
@@ -412,20 +412,22 @@ export namespace Manager.Mechanic{
     private resolveButtonDown (eventHandler: EventHandlerType, strings: string[]) {
       switch (strings[0]) {
         case 'destroy':
-          http.delete(process.env.VUE_APP_BASE_URL + 'entity' + '/' + this.id)
-            .then((response) => {
-              useToast()({
-                component: ToastComponent,
-                props: {
-                  msg: response.data.msg
-                }
-              }, {
-                type: response.data.status as TYPE
+          if (window.confirm('Are you sure you want to delete this entity?')) {
+            http.delete(process.env.VUE_APP_BASE_URL + 'entity' + '/' + this.id)
+              .then((response) => {
+                useToast()({
+                  component: ToastComponent,
+                  props: {
+                    msg: response.data.msg
+                  }
+                }, {
+                  type: response.data.status as TYPE
+                })
+                router.push({
+                  name: 'Device'
+                })
               })
-              router.push({
-                name: 'Device'
-              })
-            })
+          }
           break
         case 'codeButton':
           this.generateCode(eventHandler, strings[0])
